requirements.txt
fastapi
uvicorn
PyPDF2
python-docx
openai
python-dotenv
Jinja2
python-multipart
unittest